Fixed wrong statement about inverting noise matrix
The textbook claimed that "To obtain an inverse, we need the colums of the original matrix to be orthogonal to each other." This isn't true, you just need the determinant to be nonzero. Matrices with nonzero determinant are a set of measure zero, so you will never actually get one from any numerical procedure.

On the other hand, it IS true that when the determiant is very small, the inverse is very suceptible to small noise. This isn't fixed by using the pseudoinverse, though. You just need to do careful error propagation.