Order of qubits
I know that in qiskit qubits are numbered from right to left. Nevertheless, I think that the matrix representation of CROT_k doesn't follow this convention. To be consistent with the matrix representation, I think that the following change should be done. x_j0 ->0x_j and same for the following line.