Fix numbering in table of content, and inconsitencies in section numbering
The table of content currentl starts with the first item at 0. This is done using markdown, which works well in the ipython notebook. However, when converted into HTML, the numbering starts at 1. (not 0.) because HTML tables require a starting number as an argument. Therefore, this was changed so it starts at 1. and it is consistent between the notebook and the website.

Also, fixed section numbering. Some subsections were labeled as A., B., C., but were referenced as 4.1, 4.2, etc. in the table of contents.