Update atoms-computation.ipynb
In the half adder section. We have 4 qubits and we are counting them from 0 to 3, therefore we must use names as qubit 0, 1, 2, 3 only. However in some places of half adder "qubit 4" is used which is a bit confusing for the readers. I have updated it as "qubit 3" which is the right way to write it.