Update bernstein-vazirani.ipynb
The secret string given is actually s=011 in the python code. But it is written as s=11 in the Exercise 1 and 2. Although we can argue that 011 and 11 are equivalent in value as they both correspond to decimal value 3, but i think since it is a string then we must write s=011 in the exercise as the string 11 and 011 aren't equivalent.