Update hhl_tutorial.ipynb
I revamped this section on the HHL algorithm in "applications" to a great extent. 

I included two new, very important sub-sections i.e. Conditions and Importance in ML. The algorithm has crucial applications in ML but they were not touched upon previously. Conditions of the HHL algorithm were not explained properly either, rather briefly gone over in a few sentences. I made sure to pour light over these two sub-sections in a satisfactory manner. Furthermore, I added multiple crucial updates to both the Introduction and Mathematical background sub-sections. This was all to ensure that not much was assumed beforehand so that beginners would not have too much of a hard time catching up. The Introduction section also offers a more interesting insight into the algorithm now, to foster interest in the reader. It was quite bland before. 

Non-technical content changes are as follows:

Firstly, the title was corrected and made short and crisp, consistent with other titles of sections in the Qiskit Notebook. The title was also incorrect since no "algorithm" succeeded "HHL". HHL alone is only an abbreviation for the names of the creators. Sentences were restructured to be more grammatically correct, flowy, and easy to read. Also, better vocabulary was incorporated, and errors in the overall structure were corrected, which otherwise gave a bad impression e.g. poorly placed commas and periods. There was an incorrect fact right in the beginning, which stated that this section would demonstrate the algorithm using 5-qubits, whereas the actual number was only 4. I changed the visualisation of the circuit to mpl as well, making it more attractive and easier to understand. Lastly, the citation of the references was made consistent with each other, using Harvard.